Improved Inventory System
This is the improved version of the inventory system. Major changes
listed below:
- The UI bar for pieces is now scrollable but use a bar, will be
improved in later versions
- PieceItem and Piece Minus have been removed in favor of a general
Piece actor that has the modifiable properties PieceType and PieceIcon
- Similarly, there is also a general Gem actor with modifiable
properties GemType and GemIcon
- The interface 'Inventory' has been changed to 'Lantern Interface' and
function names have been refactored accordingly
- The functionality for moving platforms based on the pieces in the
lantern has been adapted to the improved inventory system. Previously,
the trigger box blueprints would take in a string id for the accepted
piece string ID. Now, the triggerbox blueprints take in an exposed
modifiable array of PieceTypes that represent the accepted order of
PieceTypes for that platform. This means that anyone can easily create a
triggerbox for a platform and easily define a set of pieces that make
that platform move by simply adding elements to the 'AcceptedPieceTypes'
array!